+++ AppboyProject/ios/AppboyProject/AppDelegate.m
@@ -21,14 +21,22 @@
   rootViewController.view = rootView;
   self.window.rootViewController = rootViewController;
   [self.window makeKeyAndVisible];
+  [Appboy startWithApiKey:@"d0555d14-3491-4141-a9f0-ffb83e3c2a2f"
+            inApplication:application
+        withLaunchOptions:launchOptions];
 
   // Register for user notifications
   if (floor(NSFoundationVersionNumber) > NSFoundationVersionNumber_iOS_9_x_Max) {
     UNUserNotificationCenter* center = [UNUserNotificationCenter currentNotificationCenter];
     center.delegate = self;
-    [center requestAuthorizationWithOptions:(UNAuthorizationOptionAlert | UNAuthorizationOptionSound | UNAuthorizationOptionBadge)
+    UNAuthorizationOptions options = UNAuthorizationOptionAlert | UNAuthorizationOptionSound | UNAuthorizationOptionBadge;
+    if (@available(iOS 12.0, *)) {
+      options = options | UNAuthorizationOptionProvisional;
+    }
+    [center requestAuthorizationWithOptions:(options)
                           completionHandler:^(BOOL granted, NSError * _Nullable error) {
                             NSLog(@"Permission granted.");
+                            [[Appboy sharedInstance] pushAuthorizationFromUserNotificationCenter:granted];
                           }];
     [[UIApplication sharedApplication] registerForRemoteNotifications];
   } else if (floor(NSFoundationVersionNumber) > NSFoundationVersionNumber_iOS_7_1) {
@@ -41,10 +49,6 @@
       UIRemoteNotificationTypeBadge |
       UIRemoteNotificationTypeSound)];
   }
-
-  [Appboy startWithApiKey:@"d0555d14-3491-4141-a9f0-ffb83e3c2a2f"
-            inApplication:application
-        withLaunchOptions:launchOptions];
 
   [[AppboyReactUtils sharedInstance] populateInitialUrlFromLaunchOptions:launchOptions];